+++ app/src/main/java/woowacourse/movie/MovieReservationContract.kt
@@ -11,37 +11,3 @@ interface MovieReservationContract {
         fun clickPlusNumberButton()
     }
 }
-
-class MovieReservationModel {
-    var ticketCount = 1
-        private set
-
-    fun minusTicketCount() {
-        if (ticketCount > 1) {
-            ticketCount--
-        }
-    }
-
-    fun plusTicketCount() {
-        ticketCount++
-    }
-}
-
-class MovieReservationPresenter(
-    private val view: MovieReservationContract.View,
-) : MovieReservationContract.Presenter {
-    private val model: MovieReservationModel = MovieReservationModel()
-
-    val ticketCount
-        get() = model.ticketCount
-
-    override fun clickMinusNumberButton() {
-        model.minusTicketCount()
-        view.showCurrentResultTicketCountView()
-    }
-
-    override fun clickPlusNumberButton() {
-        model.plusTicketCount()
-        view.showCurrentResultTicketCountView()
-    }
-}
