+++ app/src/main/java/woowacourse/movie/MovieReservationModel.kt
@@ -1,0 +1,16 @@
+package woowacourse.movie
+
+class MovieReservationModel {
+    var ticketCount = 1
+        private set
+
+    fun minusTicketCount() {
+        if (ticketCount > 1) {
+            ticketCount--
+        }
+    }
+
+    fun plusTicketCount() {
+        ticketCount++
+    }
+}
