+++ app/src/main/java/woowacourse/movie/ViewHolder.kt
@@ -1,0 +1,13 @@
+package woowacourse.movie
+
+import android.widget.Button
+import android.widget.ImageView
+import android.widget.TextView
+
+class ViewHolder(
+    val title: TextView,
+    val poster: ImageView,
+    val screeningDate: TextView,
+    val runningTime: TextView,
+    val movieReservationButton: Button,
+)
